requirements.txt
matplotlib==3.1.0
numpy==1.19.5
Pillow==5.2.0
tqdm==4.24.0
